rename crowdin API
